server-chatgpt/requirements.txt
fastapi 
uvicorn 
openai
python-dotenv
